Fix typos, remove unused imports, rename components, move components into hackathons directory
